feat: mark read, interup

+++ app/src/main/java/info/hermiths/chatapp/model/req/MarkReadReqBody.kt
@@ -3,7 +3,7 @@ package info.hermiths.chatapp.model.req
 import com.google.gson.annotations.SerializedName
 
 data class MarkReadReqBody(
-    val seq: Long,
+    val seq: Int,
     @SerializedName("sessionID")
     val sessionId: String,
 )
